+++ crates/pytauri-plugin-notification/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- [#47](https://github.com/WSH032/pytauri/pull/47) - refactor: use `pytauri::ImplManager` as `self::ImplNotificationExt`
+
 ## [0.1.0-beta.0]
 
 [unreleased]: https://github.com/WSH032/pytauri/tree/HEAD